Fix ordering of style assertions
If value was specified but the css property doesn't exist in enzyme, the
error would be less specific, saying the property was missing. This
could potentially make debugging harder, since the message isn't as
clear as it could be. Now, if the value is given, it checks to see if
the whole thing matches, telling you not only what property was wrong,
but the value too.