Play nicely with other custom assertions
If a user of this library already has custom assertions of the same
name, previously, this library would completely overwrite them. Now it
Attempts to wrap the existing method, falling back if `actual` is not an
enzyme wrapper. If the built-in doesn't exist, the previous behavior is
kept (assert actual is enzyme).

Please refer to issue #1 for more details.