Add smart children stringifier
Implemented smarter handling of react children. Previously, if your
component rendered any children, it'd just be shown as an ellipsis. Now
it's space-smart. If the props leave enough space, it'll attempt render
as much of the children as possible (assuming they're all primitive).
Any part of the string that doesn't fit is truncated at the furthest
space character.